% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{bluegill_palette}
\alias{bluegill_palette}
\title{A palette set to the colors of a bluegill fish}
\description{
A palette set to the colors of a bluegill fish
}
\examples{
bluegill_palette
}
\keyword{data}
